20139 - add dms deprecated classes
