+++ airflow/providers/amazon/aws/operators/dms_describe_tasks.py
@@ -16,47 +16,14 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from typing import Optional
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.operators.dms`."""
 
-from airflow.models import BaseOperator
-from airflow.providers.amazon.aws.hooks.dms import DmsHook
+import warnings
 
+from airflow.providers.amazon.aws.operators.dms import DmsDescribeTasksOperator  # noqa
 
-class DmsDescribeTasksOperator(BaseOperator):
-    """
-    Describes AWS DMS replication tasks.
-
-    :param describe_tasks_kwargs: Describe tasks command arguments
-    :type describe_tasks_kwargs: Optional[dict]
-    :param aws_conn_id: The Airflow connection used for AWS credentials.
-        If this is None or empty then the default boto3 behaviour is used. If
-        running Airflow in a distributed manner and aws_conn_id is None or
-        empty, then default boto3 configuration would be used (and must be
-        maintained on each worker node).
-    :type aws_conn_id: Optional[str]
-    """
-
-    template_fields = ('describe_tasks_kwargs',)
-    template_ext = ()
-    template_fields_renderers = {'describe_tasks_kwargs': 'json'}
-
-    def __init__(
-        self,
-        *,
-        describe_tasks_kwargs: Optional[dict] = None,
-        aws_conn_id: str = 'aws_default',
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.describe_tasks_kwargs = describe_tasks_kwargs or {}
-        self.aws_conn_id = aws_conn_id
-
-    def execute(self, context):
-        """
-        Describes AWS DMS replication tasks from Airflow
-
-        :return: Marker and list of replication tasks
-        :rtype: (Optional[str], list)
-        """
-        dms_hook = DmsHook(aws_conn_id=self.aws_conn_id)
-        return dms_hook.describe_replication_tasks(**self.describe_tasks_kwargs)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.dms`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
