+++ airflow/providers/amazon/aws/sensors/dms_task.py
@@ -16,102 +16,14 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from typing import Iterable, Optional
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.sensors.dms`."""
 
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.hooks.dms import DmsHook
-from airflow.sensors.base import BaseSensorOperator
+import warnings
 
+from airflow.providers.amazon.aws.sensors.dms import DmsTaskBaseSensor, DmsTaskCompletedSensor  # noqa
 
-class DmsTaskBaseSensor(BaseSensorOperator):
-    """
-    Contains general sensor behavior for DMS task.
-
-    Subclasses should set ``target_statuses`` and ``termination_statuses`` fields.
-
-    :param replication_task_arn: AWS DMS replication task ARN
-    :type replication_task_arn: str
-    :param aws_conn_id: aws connection to uses
-    :type aws_conn_id: str
-    :param target_statuses: the target statuses, sensor waits until
-        the task reaches any of these states
-    :type target_states: list[str]
-    :param termination_statuses: the termination statuses, sensor fails when
-        the task reaches any of these states
-    :type termination_statuses: list[str]
-    """
-
-    template_fields = ['replication_task_arn']
-    template_ext = ()
-
-    def __init__(
-        self,
-        replication_task_arn: str,
-        aws_conn_id='aws_default',
-        target_statuses: Optional[Iterable[str]] = None,
-        termination_statuses: Optional[Iterable[str]] = None,
-        *args,
-        **kwargs,
-    ):
-        super().__init__(*args, **kwargs)
-        self.aws_conn_id = aws_conn_id
-        self.replication_task_arn = replication_task_arn
-        self.target_statuses: Optional[Iterable[str]] = target_statuses
-        self.termination_statuses: Optional[Iterable[str]] = termination_statuses
-        self.hook: Optional[DmsHook] = None
-
-    def get_hook(self) -> DmsHook:
-        """Get DmsHook"""
-        if self.hook:
-            return self.hook
-
-        self.hook = DmsHook(self.aws_conn_id)
-        return self.hook
-
-    def poke(self, context):
-        status: str = self.get_hook().get_task_status(self.replication_task_arn)
-
-        if not status:
-            raise AirflowException(
-                f'Failed to read task status, task with ARN {self.replication_task_arn} not found'
-            )
-
-        self.log.info('DMS Replication task (%s) has status: %s', self.replication_task_arn, status)
-
-        if status in self.target_statuses:
-            return True
-
-        if status in self.termination_statuses:
-            raise AirflowException(f'Unexpected status: {status}')
-
-        return False
-
-
-class DmsTaskCompletedSensor(DmsTaskBaseSensor):
-    """
-    Pokes DMS task until it is completed.
-
-    .. seealso::
-        For more information on how to use this operator, take a look at the guide:
-        :ref:`howto/sensor:DmsTaskCompletedSensor`
-
-    :param replication_task_arn: AWS DMS replication task ARN
-    :type replication_task_arn: str
-    """
-
-    template_fields = ['replication_task_arn']
-    template_ext = ()
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.target_statuses = ['stopped']
-        self.termination_statuses = [
-            'creating',
-            'deleting',
-            'failed',
-            'failed-move',
-            'modifying',
-            'moving',
-            'ready',
-            'testing',
-        ]
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.dms`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
