+++ dev/provider_packages/prepare_provider_packages.py
@@ -2150,6 +2150,8 @@ KNOWN_DEPRECATED_DIRECT_IMPORTS: Set[str] = {
     'This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.ec2`.',
     "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.s3`.",
     "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.s3`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.dms`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.dms`.",
 }
