+++ docs/apache-airflow-providers-amazon/operators/dms.rst
@@ -30,13 +30,13 @@ Overview
 Airflow to AWS Database Migration Service (DMS) integration provides several operators to create and interact with
 DMS replication tasks.
 
- - :class:`~airflow.providers.amazon.aws.sensors.dms_task.DmsTaskBaseSensor`
- - :class:`~airflow.providers.amazon.aws.sensors.dms_task.DmsTaskCompletedSensor`
- - :class:`~airflow.providers.amazon.aws.operators.dms_create_task.DmsCreateTaskOperator`
- - :class:`~airflow.providers.amazon.aws.operators.dms_delete_task.DmsDeleteTaskOperator`
- - :class:`~airflow.providers.amazon.aws.operators.dms_describe_tasks.DmsDescribeTasksOperator`
- - :class:`~airflow.providers.amazon.aws.operators.dms_start_task.DmsStartTaskOperator`
- - :class:`~airflow.providers.amazon.aws.operators.dms_stop_task.DmsStopTaskOperator`
+ - :class:`~airflow.providers.amazon.aws.sensors.dms.DmsTaskBaseSensor`
+ - :class:`~airflow.providers.amazon.aws.sensors.dms.DmsTaskCompletedSensor`
+ - :class:`~airflow.providers.amazon.aws.operators.dms.DmsCreateTaskOperator`
+ - :class:`~airflow.providers.amazon.aws.operators.dms.DmsDeleteTaskOperator`
+ - :class:`~airflow.providers.amazon.aws.operators.dms.DmsDescribeTasksOperator`
+ - :class:`~airflow.providers.amazon.aws.operators.dms.DmsStartTaskOperator`
+ - :class:`~airflow.providers.amazon.aws.operators.dms.DmsStopTaskOperator`
 
 One example_dag is provided which showcases some of these operators in action.
